Add `StreamData` and `StreamMessage` type guard (#37)
Adds a `StreamMessage` and, by extension, `StreamData` type guard that we reuse throughout this package. The `StreamData` type was expanded to include numbers and arrays because these were in fact permitted. Note that the tests are completely unchanged.

There is one functional change here, which is that the streams will ignore `MessageEvent.data` values for which `Array.isArray()` returns true. Permitting this was an oversight, and I don't know of a case where this could happen in practice, but I'm nevertheless highlighting it here.