Use `addEventListener` instead of `onmessage` in WebWorkerPostMessageStream (#83)
* Use addEventListener instead of onmessage in WebWorkerPostMessageStream constructor

The use of addEventListener allows proper handling of multiple event listeners. onmesssage also does not work with LavaMoat.

* Remove test for throwing error when not run in a WebWorker

This test is now unnecessary.

* Refactor WebWorkerPostMessageStream test using mocked self

The test was trying to use globalThis.self variable, but it is not always defined in the test
environment. To fix the issue, the variable can be created as a mock object before
each test, and restored to the original value after it finishes.