BREAKING: Synchronise package with module template (April 2025) (#168)
This partially synchronises the package with the module template. It was
quite out-of-date, and because of some special configurations, it's not
as trivial to bring it up-to-date as other packages. Changes included in
this PR are:

- Build the package with `ts-bridge`.
- Update CI to the latest workflows.
- Bump Yarn to v4.
- Align scripts with the module template.

Changes not included:

- Vitest.
- I tried this, but Vitest internally uses `postMessage` as well, and
using it in tests causes issues.
- Bump to latest ESLint version.
   - This would introduce a lot more changes.

## Breaking changes

- The `package.json` now contains an `exports` field, so importing from
`'./dist` is no longer possible.
- Node.js-specific functions are now exported from
`@metamask/post-message-stream/node` instead of the default entrypoint.